Crate ic_canonical_state

Conversion, filtering and encoding of Replicated State as Canonical State.

Re-exports

pub use lazy_tree::conversion::LabelLike;

pub use visitor::Control;

pub use visitor::Visitor;

Modules

encoding
Provides canonical CBOR encoding for Canonical State tree leaves.

lazy_tree
Canonical State traversal using lazy trees, i.e. a tree with nodes that are lazily initialized if and when traversed.

size_limit_visitor
Canonical State Visitor that limits the byte size of blob leaves.

subtree_visitor
Canonical State Visitor that filters the nodes seen by a wrapped Visitor to those matchng a given pattern.

visitor
Utilities for traversing the Replicated State as if it was a state in canonical form.

Constants

CURRENT_CERTIFICATION_VERSION
The Canonical State certification version produced by this crate.

Functions

traverse
Traverses state as if it was a state in canonical form using visitor v.

traverse_partial
Traverses state partially (Wasm state excluded). It's used for quick certification that can be done every round.

Type Definitions

Label
Label applied to tree edges in the Canonical State.