Module ic_canonical_state::encoding

Provides canonical CBOR encoding for Canonical State tree leaves.

In order to maintain a stable canonical representation with support for forward and backward compatibility, mirror types are provided in the types module, as well as a suite of compatibility tests (guarding against possibly incompatible changes: field addition / removal / type changes).

The canonical CBOR representation uses the "packed" CBOR format, where field names are replaced by indices (similar to protocol buffers), for more concise encoding. This is safe to do given the use of canonical types covered by compatibility tests.

Traits

CborProxyDecoder
Allows a canonical type to act as a proxy for encoding a Rust type T as canonical CBOR, provided a TryInto<T> implementation for the canonical type.

CborProxyEncoder
Allows a canonical type to act as a proxy for encoding a Rust type T as canonical CBOR, provided an Into implementation for T.

Functions

decode_message
Decodes a RequestOrResponse from canonical CBOR representation.

decode_stream_header
Decodes a StreamHeader from canonical CBOR representation.

encode_message
Encodes a RequestOrResponse into canonical CBOR representation.

encode_metadata
Encodes a SystemMetadata into canonical CBOR representation.

encode_stream_header
Encodes a StreamHeader into canonical CBOR representation.