Trait ic_canonical_state::encoding::CborProxyEncoder

pub trait CborProxyEncoder<T> {
    fn proxy_encode(t: T) -> Result<Vec<u8>, Error>;
}

Allows a canonical type to act as a proxy for encoding a Rust type T as canonical CBOR, provided an Into implementation for T.

Required methods

fn proxy_encode(t: T) -> Result<Vec<u8>, Error>

Encodes t into a vector via this proxy.

Loading content...