Function ic_canonical_state::traverse

pub fn traverse<V: Visitor>(state: &ReplicatedState, v: V) -> V::Output

Traverses state as if it was a state in canonical form using visitor v.

By supplying different visitors, one can use traverse to serialize the state for transmitting it over the network to another replica, compute a hash tree for certification or extract a specific value.