Module ic_canonical_state::lazy_tree

Canonical State traversal using lazy trees, i.e. a tree with nodes that are lazily initialized if and when traversed.

Lazy trees allow e.g. comparing 2 Canonical States without materializing them; certified ingress history access in O(log N); and they make the algorithms on Canonical State easier to write and understand.

Modules

conversion
Conversion from ReplicatedState to LazyTree.

materialize
Support for materializing (part of) a LazyTree as a LabeledTree.

Enums

Lazy
Lazy is either a computed value or a function that knows how to compute one.

LazyTree
A tree that can lazily expand while it's being traversed.

Traits

LazyFork
The trait representing interface of a fork in the lazy tree.

Functions

blob
A helper function that constructs a leaf with a lazy blob.

follow_path
A function that extracts a value from the lazy tree by the specified path.

fork
A helper function to construct a fork of a lazy tree.

num
A helper function that construct a leaf from a number.

string
A helper function that construct a leaf from a string.

Type Definitions

ArcFn
A type alias for a ref-counted stateless function.