Module ic_canonical_state::lazy_tree::materialize

Support for materializing (part of) a LazyTree as a LabeledTree.

Functions

materialize_partial
This function provides materializing a LazyTree partially, including only the paths whose prefixes are listed in the pattern (which is itself represented as a LabeledTree, without data in the leaves).

Type Definitions

TreePattern
A pattern to be used for fitering the parts of a LazyTree to be materialized. A LabeledTree with no values.