Function ic_canonical_state::lazy_tree::materialize::materialize_partial

pub fn materialize_partial(
    lazy_tree: &LazyTree, 
    pattern: &TreePattern
) -> Option<LabeledTree<Vec<u8>>>

This function provides materializing a LazyTree partially, including only the paths whose prefixes are listed in the pattern (which is itself represented as a LabeledTree, without data in the leaves).

This is used in the implementation of the read_state request, so a specification can be found in the Interface Spec, section on Lookup in certification (https://sdk.dfinity.org/docs/interface-spec/index.html#_lookup)

The spec requires that the resulting certificate also proves the absence of a prefix; this is not yet implemented. Once it is implemented, remove the corresponding exclude directives from tests/ic-ref-test/run.