Module ic_canonical_state::visitor

Utilities for traversing the Replicated State as if it was a state in canonical form.

Enums

Control
Data structure controlling the traversal of the state tree.

Traits

Visitor
Visitor can compute a value on a canonical state tree without using an explicit intermediate representation of the tree.

Functions

named_blob
Visits the blob after entering edge with the given name.

named_num
Visits the num after entering edge with the given name.

named_subtree
Executes a function in a scope of a new subtree after entering edge with the given name.

subtree
Executes a function in the scope of a new subtree.

with_edge
Enter edge labeled as name and execute f if it was not skipped.