Function ic_canonical_state::visitor::named_subtree

pub fn named_subtree<V: Visitor, Name: AsRef<[u8]>>(
    v: &mut V, 
    name: Name, 
    f: impl FnOnce(&mut V) -> Result<(), V::Output>
) -> Result<(), V::Output>

Executes a function in a scope of a new subtree after entering edge with the given name.

The precise sequence is:

Enter an edge labeled as name on the visitor v. If the visitor returns Ok(Control::Skip), do nothing else.

Start a subtree.

Execute f on v.

Exit the subtree.