Trait ic_canonical_state::visitor::Visitor

pub trait Visitor {
    type Output;
    fn start_subtree(&mut self) -> Result<(), Self::Output>;

    fn end_subtree(&mut self) -> Result<(), Self::Output>;

    fn enter_edge(&mut self, label: &[u8]) -> Result<Control, Self::Output>;

    fn visit_num(&mut self, num: u64) -> Result<(), Self::Output>;

    fn visit_blob(&mut self, blob: &[u8]) -> Result<(), Self::Output>;

    fn finish(self) -> Self::Output;
}

Visitor can compute a value on a canonical state tree without using an explicit intermediate representation of the tree.

NOTE: once visitor returns a value, the traversal is considered complete.

Example:

Assume we want to traverse the following tree:

*
|
+-- blobmap --+-- 0 -- [ 0xCAFEBABE ]
|             |
|             +-- 1 -- [ 0xDEADBEAF ]
|
+--   num   --+-- min -- [ 5 ]
              |
              +-- max -- [ 10 ]

The traversal will produce the following trace (assuming all the calls return Ok):

visitor.start_subtree(); // root

  visitor.enter_edge(b"blobmap");
  visitor.start_subtree();

    visitor.enter_edge(b"0");
    visitor.visit_blob(&[0xCA, 0xFE, 0xBA, 0xBE]);

    visitor.enter_edge(b"1");
    visitor.visit_blob(&[0xDE, 0xAD, 0xBE, 0xEF]);

  visitor.end_subtree(); // "blobmap"

  visitor.enter_edge(b"num");
  visitor.start_subtree();

    visitor.enter_edge(b"min");
    visitor.visit_num(5);

    visitor.enter_edge(b"max");
    visitor.visit_num(10);

  visitor.end_subtree(); // "num"

visitor.end_subtree(); // root

visitor.finish();

Associated Types

type Output

If the traversal succeeds, it produces a value of this type.

Loading content...
Required methods

fn start_subtree(&mut self) -> Result<(), Self::Output>

Start a subtree traversal.

fn end_subtree(&mut self) -> Result<(), Self::Output>

Complete the subtree traversal.

fn enter_edge(&mut self, label: &[u8]) -> Result<Control, Self::Output>

Descend into the subtree along the edge labeled with label.

The visitor has a way to skip the subtree by returning Ok(Control::Skip). If that happens, none of the subtree traversal methods ((start|end)_subtree, visit_(num|blob)) will be invoked on this visitor.

For example, assume that the visitor always returns Ok(Control::Skip) from this method. This is how the trace for the example tree above will look like:

visitor.start_subtree();
  visitor.enter_edge(b"blobmap"); // => Ok(Control::Skip)
  visitor.enter_edge(b"num");     // => Ok(Control::Skip)
visitor.end_subtree();

fn visit_num(&mut self, num: u64) -> Result<(), Self::Output>

Visit a leaf of the state tree that contains a number.

fn visit_blob(&mut self, blob: &[u8]) -> Result<(), Self::Output>

Visit a leaf of the state tree that contains a blob.

fn finish(self) -> Self::Output

Complete the tree traversal.

This method is supposed to be called once the tree traversal is complete but the visitor haven't returned any value yet.

Loading content...